Change console log messages to alert format
The messages were changed to alert format so that you don't have to look at the results from the console. In addition, an if statement was added to prevent an error if the user presses the cancel button.

Information about the project's goals was added to the README.md file.