Clean up the code and add an image
The functions are called inside the eventlistener, instead of the entire code being inside it.

The images folder was created and the image was added to the page. Credit was given to the author of the image in the README file.

Other changes concerned fonts and colors on the site.